+++ CHANGELOG.md
@@ -47,7 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Inquiry
 - Inquiry Sarkome Mock-Up
 - Generate router link from window.location.pathname instead of router
-- Get export from teiler
+- Excel export from teiler
+- CSV export from teiler
 
 ## Fixed
 - Update teiler apps in sidebar
@@ -55,7 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - i18n
 - space after welcome
 - Redirect to main if router link not found
-- Copy Assets env.js in Docker 
+- Copy Assets env.js in Docker
 
 ## CHANGED
 - Check for both frontend and backend